pyproject.toml
[project]
name = "denv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "alembic>=1.16.4",
    "fastapi>=0.116.1",
    "psycopg2>=2.9.10",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]

